# openapi.yaml
openapi: 3.0.0
info:
  title: Flyway API
  description: API for Flyway application
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development server
...
components:
  schemas:
...
    BuildResponse:
      type: object
      description: Build information
      required:
        - id
        - organizationId
        - bundleId
        - commitHash
        - branchName
        - buildSize
        - buildUrl
        - nativeVersion
        - uploadedBy
        - createdAt
      properties:
        id:
          type: string
          description: Unique UUID identifier for the build
        organizationId:
          type: string
          description: Organization ID
        bundleId:
          type: string
          description: Bundle identifier for the mobile application
        commitHash:
          type: string
          description: Git commit hash (unique identifier)
        branchName:
          type: string
          description: Git branch name
        commitMessage:
          type: string
          description: Git commit message
        buildSize:
          type: integer
          format: int64
          description: Build file size in bytes
        buildUrl:
          type: string
          description: URL to download the build file
        nativeVersion:
          type: string
          description: Native app version (e.g., "1.0.0")
        uploadedBy:
          type: string
          description: User ID who uploaded the build
        createdAt:
          type: string
          format: date-time
          description: When the build was uploaded
        updatedAt:
          type: string
          format: date-time
          description: When the build was last updated
    PaginatedBuildResponse:
      type: object
      description: Paginated build response
      required:
        - data
        - page
        - size
        - totalElements
        - totalPages
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BuildResponse'
          description: Array of builds for this page
        page:
          type: integer
          description: Current page number
        size:
          type: integer
          description: Items per page
        totalElements:
          type: integer
          description: Total number of builds
        totalPages:
          type: integer
          description: Total number of pages
...
    AuditLogResponse:
      type: object
      description: Audit log entry
      required:
        - id
        - action
        - createdAt
      properties:
        id:
          type: string
          description: Unique identifier for the audit log entry
        userId:
          type: string
          description: ID of the user who performed the action
        organizationId:
          type: string
          description: ID of the organization
        action:
          type: string
          description: Action performed (e.g., MOBILE_APP_CREATED, BUILD_UPLOADED)
        resourceType:
          type: string
          description: Type of resource affected (e.g., MOBILE_APPLICATION, BUILD)
        resourceId:
          type: string
          description: ID of the affected resource
        resourceName:
          type: string
          description: Name of the affected resource
        httpMethod:
          type: string
          description: HTTP method used (GET, POST, PUT, DELETE)
        endpoint:
          type: string
          description: API endpoint called
        ipAddress:
          type: string
          description: IP address of the user
        userAgent:
          type: string
          description: User agent string
        requestBody:
          type: string
          description: Request body (if applicable)
        responseStatus:
          type: integer
          description: HTTP response status code
        errorMessage:
          type: string
          description: Error message (if action failed)
        metadata:
          type: string
          description: Additional metadata in JSON format
        createdAt:
          type: string
          format: date-time
          description: When the action occurred
    PaginatedAuditLogResponse:
      type: object
      description: Paginated audit log response
      required:
        - data
        - page
        - size
        - totalElements
        - totalPages
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogResponse'
          description: List of audit log entries
        page:
          type: integer
          description: Current page number
        size:
          type: integer
          description: Number of items per page
        totalElements:
          type: integer
          description: Total number of audit log entries
        totalPages:
          type: integer
          description: Total number of pages
    ErrorResponse:
      type: object
      description: Error response
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        details:
          type: string
          description: Error details
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
  responses:
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnauthorizedError:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /{orgId}/{bundleId}/builds:
    get:
      operationId: getBuilds
      summary: Get builds with pagination and sorting
      tags:
        - Builds
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
          description: Organization ID
        - name: bundleId
          in: path
          required: true
          schema:
            type: string
          description: Bundle ID of the mobile application
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Page number (default 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Page size (default 20)
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort direction by createdAt (default desc)
      responses:
        '200':
          description: Builds retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBuildResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    
    post:
      operationId: uploadBuild
      summary: Upload a new build
      deprecated: true
      description: "DEPRECATED: This endpoint has been removed. Use the API key-based upload endpoint at /v1/builds instead."
      tags:
        - Builds
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
          description: Organization ID
        - name: bundleId
          in: path
          required: true
          schema:
            type: string
          description: Bundle ID of the mobile application
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - commitHash
                - branchName
                - nativeVersion
                - file
              properties:
                commitHash:
                  type: string
                  description: Git commit hash (unique identifier for the build)
                branchName:
                  type: string
                  description: Git branch name
                commitMessage:
                  type: string
                  description: Git commit message
                nativeVersion:
                  type: string
                  description: Native app version (e.g., "1.0.0")
                file:
                  type: string
                  format: binary
                  description: The build file (APK/IPA, max 30MB)
      responses:
        '201':
          description: Build uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /builds/{buildId}:
    delete:
      operationId: deleteBuild
      summary: Delete a build by its UUID
      tags:
        - Builds
      security:
        - BearerAuth: []
      parameters:
        - name: buildId
          in: path
          required: true
          schema:
            type: string
          description: The UUID of the build to delete
      responses:
        '204':
          description: Build deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /v1/builds:
    post:
      operationId: uploadBuildWithApiKey
      summary: Upload a new build using API key authentication
      description: "Upload builds using API key authentication. This is the only way to upload builds and is designed for CI/CD pipelines. Requires X-API-Key header instead of JWT token."
      tags:
        - Builds
        - API Keys
      parameters:
        - name: X-API-Key
          in: header
          required: true
          schema:
            type: string
          description: API key for authentication (e.g., "flyway_abc123...")
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - commitHash
                - branchName
                - nativeVersion
                - file
              properties:
                commitHash:
                  type: string
                  description: Git commit hash (unique identifier for the build)
                branchName:
                  type: string
                  description: Git branch name
                commitMessage:
                  type: string
                  description: Git commit message
                nativeVersion:
                  type: string
                  description: Native app version (e.g., "1.0.0")
                file:
                  type: string
                  format: binary
                  description: The build file (APK/IPA, max 30MB)
      responses:
        '201':
          description: Build uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
  /audit-logs:
    get:
      operationId: getAuditLogs
      summary: Get audit logs for an organization
      description: Retrieve paginated audit logs with optional filtering by action, resource type, user, and date range
      tags:
        - Audit Logs
      security:
        - BearerAuth: []
      parameters:
        - name: action
          in: query
          required: false
          schema:
            type: string
          description: Filter by action (e.g., MOBILE_APP_CREATED)
        - name: resourceType
          in: query
          required: false
          schema:
            type: string
          description: Filter by resource type (e.g., MOBILE_APPLICATION)
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: Filter by user ID
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Filter by start date
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Filter by end date
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Page number (0-indexed)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Page size
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order by created date
      responses:
        '200':
          description: Audit logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'